requirements.txt
click==8.1.7
pyyaml==6.0.1
python-dateutil==2.8.2
pathlib-objects==1.0.2
tqdm==4.66.1
